src/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-rest-auth==7.0.1
Django==4.2.19
django-allauth==65.4.1
django-bootstrap-v5==1.0.11
django-debug-toolbar==5.0.1
django-filter==25.1
django-jazzmin==3.0.1
django-rosetta==0.10.1
django-sslserver==0.22
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.2
pillow==11.1.0
polib==1.2.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
